list comprehension camera rays

+++ src/Vector.hs
@@ -1,7 +1,7 @@
 module Vector where
 
 
-data Vector = Vector Float Float Float deriving (Eq, Show)
+data Vector = Vector Double Double Double deriving (Eq, Show)
 
 add :: Vector -> Vector -> Vector
 add (Vector ax ay az) (Vector bx by bz) = Vector (ax+bx) (ay+by) (az+bz)
@@ -9,13 +9,13 @@ add (Vector ax ay az) (Vector bx by bz) = Vector (ax+bx) (ay+by) (az+bz)
 sub :: Vector -> Vector -> Vector
 sub (Vector ax ay az) (Vector bx by bz) = Vector (ax-bx) (ay-by) (az-bz)
 
-mult :: Vector -> Float -> Vector
+mult :: Vector -> Double -> Vector
 mult (Vector x y z) n = Vector (x*n) (y*n) (z*n)
 
-mag :: Vector -> Float
+mag :: Vector -> Double
 mag v = sqrt (magSq v)
 
-magSq :: Vector -> Float
+magSq :: Vector -> Double
 magSq (Vector x y z) = x*x + y*y + z*z
 
 normalize :: Vector -> Vector
